requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==19.3.0
audioread==3.0.1
Automat==0.8.0
Babel==2.13.1
backcall==0.2.0
bce-python-sdk==0.8.95
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
bokeh==3.1.1
boltons==23.1.1
Bottleneck==1.3.7
braceexpand==0.1.7
bs4==0.0.1
cached-property==1.5.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
ci-info==0.3.0
click==8.1.7
cloud-init==23.1.2
cloudpathlib==0.16.0
cmdtest==0.32+git
colorama==0.4.3
coloredlogs==15.0.1
colorlog==6.7.0
command-not-found==0.3
confection==0.1.3
configobj==5.0.6
configparser==6.0.0
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
cymem==2.0.8
Cython==3.0.5
datasets==2.14.6
dbus-python==1.2.16
decorator==4.4.2
dill==0.3.4
Distance==0.1.3
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
editdistance==0.6.2
einops==0.7.0
entrypoints==0.3
etelemetry==0.3.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.68.1
ffmpy==0.3.1
filelock==3.13.0
Flask==3.0.0
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.44.0
frontend==0.0.3
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
g2p-en==2.1.0
g2pM==0.1.2.5
gradio==3.35.0
gradio_client==0.7.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.1
httplib2==0.14.0
httpx==0.25.1
huggingface-hub==0.18.0
humanfriendly==10.0
hyperlink==19.0.0
HyperPyYAML==1.2.2
idna==2.8
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==16.10.1
inflect==7.0.0
intervaltree==3.1.0
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kaldiio==2.18.0
keyring==18.0.1
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.3
Levenshtein==0.23.0
librosa==0.8.1
linkify-it-py==2.0.2
llvmlite==0.41.1
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
Markdown==3.1.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mido==1.3.0
mock==5.1.0
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.12.2
murmurhash==1.0.10
nara-wpe==0.0.9
netifaces==0.10.4
networkx==3.1
nftables==0.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
note-seq==0.0.3
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.1
openai==0.28.1
OpenCC==1.1.7
opencc-python-reimplemented==0.1.7
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
paddle2onnx==1.1.0
paddleaudio==1.1.0
paddlefsl==1.1.0
paddlenlp==2.6.1
paddlepaddle==2.5.2
paddlesde==0.2.5
paddleslim==2.4.1
paddlespeech==1.4.1
paddlespeech-feat==0.1.0
pandas==2.0.3
parameterized==0.9.0
parso==0.8.3
pathlib==1.0.1
pathos==0.2.8
pathy==0.10.3
pattern-singleton==1.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==3.11.0
pooch==1.8.0
portalocker==2.8.2
pox==0.3.3
ppdiffusers==0.19.3
ppft==1.7.6.7
praatio==5.1.1
preshed==3.0.9
pretty-midi==0.2.10
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==3.20.2
prov==2.0.0
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydantic_core==2.10.1
pydot==1.4.2
pydub==0.25.1
Pygments==2.16.1
PyGObject==3.36.0
pygtrie==2.5.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyNaCl==1.3.0
pynvml==11.5.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypinyin==0.44.0
pypinyin-dict==0.6.0
pyrsistent==0.15.5
pyserial==3.4
pytest-runner==6.0.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-multipart==0.0.5
pytz==2023.3.post1
PyWavelets==1.4.1
pyworld==0.3.4
pyxnat==1.6
PyYAML==5.3.1
pyzmq==25.1.1
qudida==0.0.4
rapidfuzz==3.4.0
rarfile==4.1
rdflib==7.0.0
regex==2023.10.3
requests==2.26.0
requests-mock==1.11.0
requests-unixsocket==0.2.0
resampy==0.4.2
retrying==1.3.4
rich==13.6.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sacrebleu==2.3.2
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
selinux==3.0
semantic-version==2.10.0
sentencepiece==0.1.99
seqeval==1.2.2
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
slip==0.6.5
slip.dbus==0.6.5
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.4
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.10
stack-data==0.6.3
starlette==0.14.2
swig==4.1.1.post0
sympy==1.12
systemd-python==234
tabulate==0.9.0
TextGrid==1.5
texttable==1.6.2
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.5.1
timer==0.2.2
timm==0.5.4
ToJyutping==0.2.3
tokenizers==0.13.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
traits==6.3.2
trampoline==0.1.2
transformers==4.27.1
triton==2.1.0
ttystatus==0.38
Twisted==18.9.0
typeguard==2.13.3
typer==0.4.2
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
uc-micro-py==1.0.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.15.0
visualdl==2.5.3
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.9
weasel==0.3.3
webrtcvad==2.0.10
websocket-client==0.53.0
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
xyzservices==2023.10.1
yacs==0.1.8
yarg==0.1.9
yarl==1.9.2
zhon==2.0.2
zipp==3.17.0
zope.interface==4.7.1
